Add functionality to manually generate mipmap level from IBitmapView
Add functionality to manually generate mipmap level from IBitmapView.
The generation will halve the resolution of the source bitmap on each
axis and combine blocks of 2x2 pixels into a single pixel in the new
generated bitmap.

sRGB encoded RGB or RGBA bitmaps are supported and will be correctly
converted into linear for filtering and then back into sRGB.

Source bitmaps with odd number of rows or columns will be clamped
to to an even number and the pixel information from the odd lines
will be discarded and not contribute to the final output.

The intended use case is for manually generating mip map levels
when glGenerateMipmap cannot be used. For example when using sRGB
extension as texture type with GL ES2.