Add a visual sRGB rendering test case in the graphics test app
Add a visual sRGB rendering test case in the graphics test app. The
goal of this test is to be able to get some clues whether the sRGB
based rendering and color buffer encoding is working or not. There
are many caveats, more details are in the associated comment in the
test case code.