Separate different texture types to different texture packs when packing
Separate different texture types to different texture packs when packing
so that RGBA, RGB and Grayscale (Alpha mask) textures all get packed
into texture packs/atlasses that have the same bit depth as the source
images do. In other words, don't pack alpha mask textures in RGB(A)
texture packs or vice versa.

QImage seems to lie about the source image depth so as a workaround
for now we're using the gfx::Image to load the content to check what
is the actual source image bit depth and then load it again (ugh)
in the QImage for resampling purposes.