Add first WIP graphics support for pre-multiplied alpha textures
Add first WIP graphics support for pre-multiplied alpha textures.

The fundamental problem is that with alpha mask/cutout textures
filtering will mix transparent pixels with the actual pixels at
the frindges of the cutout object. This will create artifacts (so
called fringing) when blending the output from the fragment shader
with the previous contents of the color buffer.

The solution is to use so called "pre-multiplied" alpha where the
alpha values are baked into the RGB values and the blending equation
then no longer modulates the src RGB based on src alpha but only uses
the src alpha to modulate the dst.

Overall this functionality requires:
- a texture flag to indicate whether the texture has (or should have)
  alpha pre-multiplied or not
- a texture pre-load pass that performs the actual alpha premultiplication
  on texture load
- a material "pre-multiplied alpha" flag so that the blending  equation
  can change from properly when flag is set

TODO:
- alpha premult can be done when packing.
- need to check the effect of using this with the built-in shaders
  for sprites and texture2Ds when using the color modulation.