Add scripting variable array support
Add scripting variable array support. This commit includes vertical
changes to support array type of a scripting variable. The array
will be accessed in Lua through a ArrayInterface object which provides
the Lua binding interface (with some additional methods) for accessing
the underlying scripting variable array data. If the object is not
array then the Lua object that is returned on variable access is
the primitive object.