Refactor uikit widget action polling semantics and API
Refactor uikit widget action polling semantics and API. The new
version of this functionality will poll all widgets for actions
instead of just the currently active widget and then return a
vector of actions instead of just a single action. This change
allows us to have a widget generate an action as a response to some
other event/method call other than a mouse/keyboard event.

For example a radio button widget might generate an action event
when its state is changed programmatically in order to properly
implement the exclusive selection constraint (within a group of
radio buttons).