Add first version of a radio button widget to the uikit
Add first version of a radio button widget to the uikit. Some design
time features such as copy/pasting or deleting radio button widgets
don't properly implement design time state setting and making sure
that exlusive radio button selection is maintained within a radio
button group. These are left for later commits if needed.

When the uikit client application wants to programmatically set
a radio button as the selected button within a radio button group
there's a need to maintain the exclusive radio button property.
I.e. all the other buttons in that group should then be turned off.
To that extent the radio button stores an action that will be received
exactly once when calling PollAction. Upon receiving this action the
window will then find the sibling radio buttons and set the selection
flag appropriately. Howevef there's a small chance of a bug happening
depending exactly how and when if Widget::PollAction is called and
in which order and when wrt mouse/keyboard event processing. For
example:

- game calls Select on some radio button A
-  the action gets stored in the radio button and not immediately
   processed
- mouse input event arrives clicking on some other radio button B
-  the radio button processes the event and stores a new action
- window polls the radio buttons for actions
-  if button B is polled first and button A later the result is that
   A is selected even though B is the expected result.

This is the result of essentially using an action queue (with just 1 item)
in each radio button instead of using an action queue per radio button
group. The reason this design is used is because using an action queue
per radio button group would be much more complicated to implement.