Add possibility to add inline widget styling to a window in uikit
Add possibility to add inline widget styling to a window in uikit.
This defines an additional level of styling information that can be
applied and that would be used between widget specific styling and
generic base level styling. The intended feature is that the the
window can define styling information that overrides some particular
styling coming from the style file and which applies to all widgets
of some particular type placed in this window. Then widget specific
styling can then be futher used to (re)define the actual look and
appearance of any specific widget.