Check against minimum size required in spinbox widget in uikit
Check against minimum size required in spinbox widget in uikit. The
code srinks the widget's box by some amount in order to create a rect
for the editable text. If the widget's rect is smaller by the growth
amount in either dimension we get a negative value which in the text
rasterizer gets turned into a large unsigned value which is incorrect.

The fix is to simply check the spinbox's rect size against a minimum
rect size that will avoid this problem.