Refactor program projection/view/model matrix setting into drawable
Refactor program projection/view/model matrix setting into drawable.
This will allow the vertex shader implementation to use different
chains of transformations without the painter having to know about
the actual uniforms. For example in the future when the particle
engine can operate with particles in the global coordinate space the
vertex shader needs to transform the vertices with view and projection
matrices only and model transform can be omitted.