Update perf test with some small changes and a new render test case
Update perf test with some small changes and a new render test case.
The new render test case uses a more complicated entity with several
nodes in the entity's render tree. This test case increases the CPU
computation in order to have a more realistic view where the time
will be spent in the rendering and which parts to optimize.