Fix initial original hash values in various editor widgets
Fix initial original hash values in various editor widgets. The hash
values need to be initialized to whatever is the hash of the object
being edited so that when closing the widget without any changes being
made the current checks out to be the same as the actual hash value
of the object under edit. This avoids asking for close confirmation
when no actual changes have been done.