Refactor widget close confirmation into mainwindow and childwindow
Refactor widget close confirmation into mainwindow and childwindow.
There's no need to repeat this functionality in every single widget
since a function already exists to check for unsaved changes. The
dialog to ask for confirmation to close the widget with unsaved changes
can then be triggered from the main and child windows that manage the
actual widget.