Improve pairing of vertex and fragment programs
Improve pairing of vertex and fragment programs. Instead of creating
a new program object for every pair of shape/material it's enough
to create a program object for every vertex-shader/material pair.

For example two shapes, a rect and a circle, can both be drawn
using the same program when they're both using the same material.
We'd only need to create different programs if the shapes mapped
to different vertex shaders.

By adding a proper implementation for the GetProgramId in the drawable
interface we can map the shape to a more reasonable ID based on the
what is the actual shader program.