Add support for particle simulation in global coordinate space
Add support for particle simulation in global coordinate space.
Previously all particles were always simulated relative to a local
coordinate space. This has the problem that when an entity node that
has a particle system as a drawable is moved all the particles are
moved with it and this can sometimes not be desired. For example if
you imagine a space ship moving in space it might want to leave a
trail of particles behind it as it moves. Yet with local space all
the particles always move as the space ship moves.

The solution to this problem is letting particles be simulated in
global coordinate space. In this mode the particles are spawned
directly using world space coordinates and are updated in world space.
In order to make this happen we use the model matrix in the drawbable
environment object to figure out what are the world extents and the
position of the drawable node. These information are then used to
spawn the particles within some area in the world.

Since the particle position is directly in world coordinates then
rendering transform in vertex shader doesn't need to multiply the
vertices by model and view transform we can provide an alternative
shader implementation which only transforms by view matrix.

A yet unsolved issue is culling the particles against some boundary
values. This is currently left unsolved and will need to be adddressed
in some later commit(s).