Allow for some difference in the renderer unit tests
Allow for some difference in the renderer unit tests between the
rendered output and the expected result. Commit 2cbc97b5 changed the
behaviour of the renderer and this now triggers a problem where the
rendered shapes are not on exact pixel boundaries. The tests originally
expect exact mapping from game/world space coordinates to certain pixel
locations in the render target and this expectation is too stringent.
Add a new testing system that allows for some error treshold when
counting the number of pixels with certain in some region of the
render target.