Pass normalized particle time to the material shader
Pass normalized particle time to the material shader. Use the configured
maximum particle lifetime to derive a normalized particle lifetime
value and pass it from the vertex shader to the material shader.
This is convenient for example interpolating between two colors in
the material shader as a particle is fading from one color to another
as it expires.