Alternate with border and borderless when drawing background grid
Alternate with border and borderless when drawing background grid
in order to avoid a rendering artifact in the middle when two grids
with borders overlap and are drawn semi transparent. Blending a line
two times over at the same time over creates an unwanted "thickening"
or "higlight" effect at those fragments. If we draw only one grid
at both halves with border lines then the problem is solved.