Add min/max particle simulation duration and one off delay at the start
Add min/max particle simulation duration and one off delay at the start.
When combining multiple particle emitters for one effect there's a
need to have particular effect happen at a specific time in order for
the effect to render the desired visual output. For example with an
explosion the smoke right might appear after the flash. It'd be
possible to combine the emitters into an entity and use drawable time
scaling and entity animation tracks but providing simple time controls
with the particle simulation directly simplifies things.