Fix bugs related to min/max setting in the particle widget
Fix bugs related to min/max setting in the particle widget. As
expected Qt can bite one in the ass by triggering a signal when
min/max is set if the value is asjusted resulting in unwanted
recursion. Yay!

Use the parameter object for setting the spinbox min/max values
and create a new wrapper function for setting the ranges while
blocking the signals.