Refactor how texture maps are applied when reading shader JSON
Refactor how texture maps are applied when reading shader JSON.
If there's a previous texture map with a matching type  use that
so that the old texture sources are preserved and simply update the
sampler and texture rectangle names.

This also fixes a bug where a change in texture rect uniform name
would not be reflected in the material when applying shader shader
uniforms but instead the old (latent) name would be used.