Refactor shader file loading from device into material
Refactor shader file loading from device into material. This is a
more appropriate place for this. Additionally fix the custom material
program ID checking. The ID can/should be derived from the shader
source or the source file URI so that multiple classes with the same
underlying GLSL file will map to the same shader object.