Rename base::RenderTree template private functions to avoid weird issues
Rename base::RenderTree template private functions to avoid weird issues
such as a call to ForEachChild failing because it's "private" because the
call resolution resolved the call (for whatever weird reason) to the
private template overload instead of the public overload. Don't have
interest to deal with this s*t so let's just rename methods in order
to avoid this sillyness.