Infer text rect dimensions from raster buffer size when not specified
Infer text rect dimensions from raster buffer size when not specified.
This unfortunately requires that the text is infact rasterized once
in order to figure out what are the dimensions of the text block.
Currently this code path is not frequently used though so can't be
bothered to deal with proper font metrics right now.