Fix uniform accumulation bug when drawing multiple times
Fix uniform accumulation bug when drawing multiple times with
a single program object. The program object defers the actual
uniform setting until the program is used to draw and the
pending uniforms that are not yet flushed onto the GPU program
object are stored in a vector. The problem is that if the same
CPU side program object is used to draw multiple times within a
single frame all the calls to set program uniforms will accumulate
incorrectly. Instead the uniforms should be flushed onto the GPU
program object on every draw and then cleared.