Refactor renderer for alternative rendering path
Refactor renderer for alternative rendering path that is used when
running the game "for real". This rendering path leverages the
existing functionality as much as possible but provides for better
ways to later on extend the functionality in terms of how to the
scene is mapped to rendering nodes and also how to integrate the
tile rendering in the same generic update/draw mechanism. The old
ways of drawing are kept in order to support the existing drawing
functionality that is essential for the editor.

Update demo content with latest changes in engine. Fix some layering
issues in scenes caused by not having right layer indices in place.