Fix awesome Qt JSON/Variant derp again
Fix awesome Qt JSON/Variant derp again. On x86-64 size_t is 64bits
which silently goes into QVariant which goes into QJsonDocument
but silently borks.

Fixing this bug also fixes several bugs regarding the state of the
mainwidgets and whether the system thinks there are changes that
need to be saved. These bugs happened because of using size_t for
hashing and the hash value would not be correctly restored on
session load. Barf.

Add a unit test for testing some more of the code in the editor/gui
in the future.