Fix too optimistic optimization in sparse tilemap cache save
Fix too optimistic optimization in sparse tilemap cache save. The
tile write cannot be omitted when the new tile value is the same
as the default tile value if the tile has been written a different
value prior. The tile write can be omitted only if it's known that
the tile has no prior value and the simplest way to check for this
is to check whether the block exists or not.