Add an attempted workaround for Qt/X11 mouse cursor issue
Add an attempted workaround for Qt/X11 mouse cursor issue. The
issue is that occasionally the mouse cursor changes by itself to
an IBeam without that change happening in the editor. I suspect this
is a bug related to the QWindow and X11 and how the window is
embedded in the Qt widget system.

As a workaround trying to set the cursor explicitly on widget init.