Refactor tilemap layer data association and loading
Refactor tilemap layer data association and loading.
- Add a way to associate the layer with an arbitrary "data object ID".
- Refactor the loader interface to use a struct of loader args.

Technically adding a data object ID is redundant with the URI but
this allows for more flexibility in terms of choosing what is the
most convenient way to discover/load the associated data file or
the resource that is associated with the data object.