Refactor workspace resource packing
Refactor workspace resource packing by creating a new inteface for
copying the data files and by moving the packing implementation
"sort of" under resource implementations.
The main goal of this change is to be able to re-use this functionality
later on with resource export/import since that functionality will
also need to essentially iterate over the resources' file based
resources and copy them over.