Add first version of proper resource export/import in GSEditor
Add first version of proper resource export/import in GSEditor.
The meaning of proper here is that the resources are collected and
combined into a zip file which can then be unzipped and contents
extracted and imported into another workspace. All dependant resources
such as shaders, font files, Lua scripts etc. are collected and
their URIs are mapped appropriately.

TODO:
- Realize and deal with errors better.
- Figure out what to do with errors, should the export/import process
  stop on errors or continue.
- Possibly cleanup intermediate files when extracting/importing and
  there's an error.
- Handle filename collisions both when zipping and unzipping.
- Skip duplicate work when for example rewriting the UI style JSON.
- User interface for initiating export/import feature.
- Handle duplicate resources, for example when trying to import
  the same package twice.