Add first version of --viewer mode in GSEditor
Add first version of --viewer mode in GSEditor. This flag enables
a view only mode in GSEditor which is essentially a limited version
of the editor itself with all editing capabilities disabled. The
goal of this is to be able to launch the viewer from some GSEditor
instance, open another workspace/project,browse the content and be
able to export/import those contents into the current workspace/project
in the GSeditor.

The reason why this is a separate process instead of just being a
dialog in the GSEditor is because essentially right now there are
some (unfortunately bad decisions) global data that is tied to the
current workspace (QAbstractFileEngineHandler and gfx::Loader) thus
basically making it impossible to have 2 separate workspace objects
at the same time. It'd require some refactoring to remove these
limitations so for the time being a simpler solution is simply to
use another process and communicate over IPC.