Add engine data loading by ID and refactor Lua script loading
Add engine data loading by ID and refactor Lua script loading. All
Lua scripts (in uik::Window, Entity and Scene) are identified by
an ID instead of using an URI. The Lua integration assumed that all
Lua scripts are in some particular location relative to the some
base path and thus used the script ID to try to conjure the actual
Lua script file path. This fails for cases where the actual file is
deployed to some other path/location than the assumed lua/ folder.

To fix this problem this commit introduces a way to load arbitrary
engine data by ID. The implementations must then map the ID to a
file somehow. In workspace this is done with the Resource objects
that are used to represent things such as scripts and data objects
and a lookup is performed by matching the incoming ID to a resource
object which then contains the actual file URI.

In the deployed game loader it's possible to skip the object creation
and simply have a ID->URI map which is loaded/initialized from the
content.json on initial game load.