Add yet another way to modify uikit widget materials
Add yet another way to modify uikit widget materials. Similar to
the commit b2f9bfe674a00753ca0649dffde03010a2ebfda6 this change
provides a method for the uikit user/client to alter some widget
materials at runtime. The new widget APIs let the user to specify
a material string that is then used by the painter to create the
actual material used to paint the object by that key.

The supported materials are simple color, gradient and a reference
to material class defined in the workspace/project.