Add widget tab order to uikit widgets and tab order UI to GSEditor
Add widget tab order to uikit widgets and tab order UI to GSEditor.
The widget tab order will be the order in which the widgets will
be traversed when keyboard navigation is supported. This commit
adds the widget tab order and the relevant UI for re-arranging the
widgets in a tab order list, visually indicating the current
tab order the UI widget in GSEditor and also a migration path for
UI resources to have the initial tab order generated when the prior
data doesn't yet have that.