Add an audio engine flag to control audio effect playback
Add an audio engine flag to control audio effect playback. This
flag is added in the audio engine simply for the large utility
value that having it here allows and letting the actual game code
remain super simple without having to check everywhere whether
audio effects are enabled or not.