Add UI functionality to use images directly in uikit widget styling
Add UI functionality to use images directly in uikit widget styling.
This change allows a quicker way to style a uikit widget with an
image by letting the user to choose an image file directly and without
having to create a material out using the image first.

The current caveats however are that:
- There's no way to change the texture filtering or any other
  texture properties.However, if this is needed then the user can go
  the material route instead.
- There's no way to combine the UI textures yet so the this is likely
  going to create a whole bunch of (small) texture objects on the GPU.