Add first version of texture atlas viewer dialog to GSEditor
Add first version of texture atlas viewer dialog to GSEditor. This
is able to display the contents of packged texture image (atlas)
accompanied by a JSON descriptor file. The format is compatible
with the JSON produced by the image packing system already existing
in the editor.

The intended use case is to be able to leverage this dialog when
choosing an image (to be used as a texture) from a image pack without
wanting to use the material system. For example a uikit UI style
might package button icons etc. outside the editor material/content
workflow once into a single image file and let the widgets use
images from the atlas. In some cases the user might want to choose
the image from the pack in which case an atlas viewer is needed.