Add possibility to select a texture atlas for a UI style texture
Add possibility to select a texture atlas for a UI style texture.
If the image file chosen as a UI texture has an associated JSON
file we'll open a dialog that lets the user to choose a specific
iamge inside the packed image file. The details of the image
selection are then stored (in addition to the image URI) in the
UITexture object.

Note that the link between UITexture and the image inside the
packed image is based on a "name" instead of any kind of ID.
This let's the image to be re-packed while maintaining the UITexture
associations that already exist. Generating IDs for the packed
images would break this association unless the IDs somehow maintained
so that every image would always get the same ID when the image
pack is re-created. This would be rather complicated. Using a simple
name derived from the source image name is simpler but there does
exist the problem of potential name collisions. But the current
solution is to let the user make sure that this doesn't happen.