Add first version of bitmap font support in graphics subsystem
Add first version of bitmap font support in graphics subsystem. This
change adds a way to use a bitmap based font in the graphics subsystem
to render text. The assumption is that the font is comprised of two files:
  1. Image file (such as a png) that contains the packed pre-rasterized
     font glyphs in RGBA format at some target resolution such as 64x64px.
  2. A descriptor file (JSON) that contains the relevant information for
     mapping Unicode characters to the glyphs.

Since the assumption is that the glyph image is a RGBA (32bit depth) the
text rasterization does not use CPU or CPU side bitmap/memory buffer but
instead relies on a GFX device provided texture object. The algorithm will
allocate a FBO and a composition texture (color render buffer target) and
use the GPU to render a geometry while sampling from the font glyph texture
when rasterizing. The resulting output is then a texture object. This also
provides a fast scaling in case the desired  "line height" doesn't match
the native pixel width/height of the glyphs.

Currently there's no provision for kerning or understanding glyph offsets
regarding baseline. Instead every glyph is assumed to have same fixed width
and height and there are no x/y offsets being applied when placing the
glyphs on the baseline.

Caveat about the implementation The FBO renders "upside down", i.e. the
the last scanline is at texture coordinate y=0.0 and the top scanline
is at y=1.0. This is important to keep in mind and currently requires
a fragment shader transformation to convert the texture coordinates.
The drawable system has flipped the texture coordinates based on the
regular use case of uploading texture images from file.

Another caveat is that the bitmap+json based font doesn't work as a
data source for material's since the rasteriation is on the GPU. This
could be fixed later with some more complications in the texture source
API.

Some potential TODO features for later time.

  * Support alpha mask texture, i.e. when the glyph texture is actually
    just an alpha mask. The FBO render target color buffer must be an RGBA
    texture thus the output will always be RGBA texture instead of one
    channel grayscale. The shader object (in TextMaterial) needs to understand
    this and realize that while the texture has multiple channels only alpha
    should be used. The alpha should then be with a color uniform to render.
  * Add possible kerning information. Some bitmap fonts such as angelcode's

    bitmap font generator can provide kerning too.

  * Add a simple x/y advance scaler for uniformly adjusting the x/y advance
    when laying the glyphs on the baseline.

  * Add character set information. Using simple ASCII might offer some
    shortcuts for better performance but this requires knowing that the
    character set is in fact just ASCII and not larger Unicode.