Add first version of bitmap font map tool in GSEditor
Add first version of bitmap font map tool in GSEditor. The use case
for this tool is to be able to take 3rd party pre-rendered bitmap
font file and create a meaningful font description (JSON) file out
of it. The resulting JSON (and the image file) can then be used to
render text using the bitmap text path in the graphics text rendering.