Add a hacky way to copy (export/import) bitmap font image files
Add a hacky way to copy (export/import) bitmap font image files.
This implementation is not robust and would need to be fixed later.

The current failure cases include:
- The font image file extension is not ".png" (casing applies).
- The font image file name doesn't match the .json file basename.

More robust implementation would require probing the .json file and
checking the filename mentioned there. A complete alternative would
be to pack the font image in the .json file as a base64 encoded string
and avoid the problem entirely.