Add missing call to delete UI style material instances in UI editor
Add missing call to delete UI style material instanced in UI editor
when textures need to be reloaded. There can be  material instances
that are using data from packed images and the texture coordinates
might have changed. When the material instances are deleted the JSON
descriptor for the packed images are re-read and any change in the
texture coordinates are realized.