Add material actuator UI to animation widget using material params dialog
Add material actuator UI to animation widget using material params dialog.
Refactor the dialog for setting material parameters a little bit so
that it can be used both for setting the entity node's drawable item's
material parameters and also for adjusting the parameters set on a
material actuator. The UX is for adjusting the material actuator
is not perfect since it requires using this extra dialog but the
dialog has the functionality already there for parsing custom shader
descriptor (.json) files and building the right UI so we're going
with that for now.

Other limitation is that currently there's no way to get the default
material uniform value at runtime so the material actuator cannot
do uniform interpolation *unless* the same parameter is changed to
some value in the entity node. I.e. the entity node has a parameter
that overrides the defaults in the material.