Refactor scene and entity render hooks and scene node selection
Refactor scene and entity render hooks and scene node selection.
Tidy up the code and move the actual drawing code to drawing.cpp.
This simplifies the logic in the draw hook and puts all the drawing
stuff in a single place. Also since this commit changes the scene
node visualization the node selection must also be updated accordingly.

There are still a few bugs regarding hitting the right tool hotspot
when theres a scaling factor in the view transformation. This needs
to be fixed later on. Also the visual indication for scene node selection
takes now place in the scene node hooks and while this makes the
drawing logic simpler the problem is that the selection drawing is now
in the scene layering and thus can be obscured by other draws. Fixing
this would require more changes in the render hook APIs.