Fix potential nullptr access on deleted entity class reference
Fix potential nullptr access on deleted entity class reference in
scene widget. When an entity class is used any scene nodes that
refer to it are updated and the entity class reference is removed.
This means that the entity class refernce is a null pointer and
subsequently any place that accessess the entity class must check
for the nullptr. Ugh!

Add a error warning print in the scene paint to indicate visually
on screen any scene nodes that refer to entity classes that no
longer exist.