Refactor all class type JSON loading interfaces
Refactor all class type JSON loading interfaces such as SceneNodeClass,
SceneClass, MaterialClass, EntityClass, etc.
The problem is that bailing on first error and returning a nullopt
class object is too restrictive and not resilient in the face of
changes in the data. It seems a better strategy is to load as much
of the data as possible (keep track of failures somewhere, currently
in the log) and return some value that indicates whether the loading
was completely successful or not. This change lets the client/caller
then to decide whether the data should be discarded or kept around.

Add logging for data load failures when reasonable. Also Properly
return the success value from several FromJson load functions.