Enhance stencil masking in game renderer
Enhance stencil masking in game renderer. Add stencil mask for the
inverse i.e. expose and allow this to be combined with the cover mask.
The net effect is that if you put a cover and an expose mask together
only the area covered by expose mask is actually draw into.