Add render pass hook for modifying the material shader source
Add render pass hook for modifying the material shader source. The
problem is that there could be different render passes with PP
effects such as bloom which will need to do some special computing
for example based on the RGB fragment value the "normal" material
shader has computed. In order to combine these two some kind of a
hook is needed in the regular shader code for invoking the specific
code that the render pass needs. This is implemented by having a

 - GLSL function declaration for a RenderPass function.
 - Calling RenderPass for creating the ultimate gl_FragColor value.
 - Having the render pass object modify the GLSL source in order to
   append the RenderPass function implementation.