Add first version of a "simple" collapsible Qt5 widget
Add first version of a "simple" collapsible Qt5 widget. The use
case for this is for replacing the tabbed widget with a better
looking alternative where the tab widget doesn't work/look nice.
Simple idea, have a container widget that allows to be closed or
expanded.

Seems that writing the Qt5 UI Designer plugin for a container widget
is a mix of rocket science and black arts. The container plugin
example specifically mentions that the interface is for "page widgets"
but in practice this seems to be only possible way to do this kind
of widget. I.e. using a paged stacked widget with only a single page.

Second problem is to do with the tab focusing. This is currently
likely to be broken in some subtle ways. Again rocket science meets
black arts. The Qt5 documentation is junk (I suspect it has literal
lies in it) but looking at some Qt widget sources it at least looks
like the focusNextPrevChild should return false when there was no
next widget and true if there was (and setFocus on it).