Add a workaround for some QWindow Qt bugs in editor's PlayWindow
Add a workaround for some QWindow Qt bugs in editor's PlayWindow.
The problem is that doing something like raise() or activateWindow()
on the QWindow that is used as the OpenGL rendering surface in the
PlayWindow (which is a QMainWindow) often do not work as expected.
It seems that there's a need to do some kind of a delay after the
main window has been shown and before these calls can be done so
that the result is what is expected.