Add first version of "preview" feature in entity widget
Add first version of "preview" feature in entity widget.
The motivation is to be able to do "more" than is possible in the
editor during the design state but without having to launch the whole
game. For example it might be desireable to see how entity animations
look or how it reacts to key presses or other input. This is not
possible in the editor since there'd be too much possible state,
the editor would be confusing if things moved around in the entity
designer widget. Instead it's possible to leverage the existing
PlayWindow for running the whole game but instead of lauching with
the normal game "main" script we can specify an alternative script
that can take shortcuts (or do other specific stuff) to set the state
just right for previewing the entity.