Refactor graphics device/context interface and device creation
Refactor graphics device/context interface and device creation.
Move the current OpenGL ES2 device implementation under device/
and create a new device interface. The goal of this change is to
prepare the architecture for the addition of a new interface for
compute that may or may not be the same as the graphics device.
The new device creation interface allows for the creation of a
generic GPU device which can then be created for an access to a
graphics or compute device specifically.