Refactor gfx::RenderPass into two classes
Refactor gfx::RenderPass into two classes. A lower layer ShaderPass
which can be used to modify the shader source and add additional
or modify original program uniforms. Add a higher level RenderPass
class which takes the painter and sets the FBO state on every draw
call to the painter. This change is needed so that it's possible to
do multiple low level draws inside a high level render pass. For
example we might want to render to an offscreen buffer for PP effects
while doing stencil mask and stencil tests.