Add first version of bloom filter to whole scene rendering
Add first version of bloom filter to whole scene rendering. The
implementation mimics a "real HDR" bloom filter in the sense that it
tries to extract a "brightness" (lightness?) value of the rendered
color value and then comparing that to a threshold value and then
writes these values to a texture which is then blurred and blended
(additive) on top of the scene again. The limitations of of GL ES2
require some inefficiency in the implementation. Essentially:

- There's no universal MSAA support for FBOs. (There's of course
  EXT_multisampled_render_to_texture but there's no WebGL support).
- There's no multiple render target/color buffer support, i.e. it's
  not possible for the shader to write to multiple FBO attachments
  on a single shader pass.

Other possible things to do:
- Entity flags for turning off the contribution to bloom per entity.
- Lua interfaces for the renderer to control PP effects (in general).