Refactor texture uploading in texture maps and texture sources
Refactor texture uploading in texture maps and texture sources. This
change adds a new higher level function for uploading textures onto
the graphics device and returning a gfx::Texture* object. The onus
of doing the actual upload is moved into the texture source class
implementations. This simplifies the texture maps (both 2D and sprite)
and allows the texture sources more flexibility in how the GPU texture
is sourced. This in turns allows the use of texture objects that are
generated by some other way and already exist on the device such as
bitmap text rasterized with the GPU. Also it's now possible to remove
some TextureSource methods that are no longer needed such as
GetContentHash and GetGpuId. Note that the GetData is still left in
the interface because it's used to for example obtain a preview
copy of the texture data in the Editor.